requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==3.7.1
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
courlan==1.3.2
dateparser==1.2.1
distro==1.9.0
dotenv==0.9.9
Flask==2.3.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
gunicorn==21.2.0
h11==0.14.0
htmldate==1.9.3
httpcore==0.18.0
httpx==0.25.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
MarkupSafe==3.0.2
msgspec==0.19.0
openai==1.3.7
packaging==25.0
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uuid==1.30
waitress==3.0.2
Werkzeug==3.1.3
openai==1.3.7
httpx==0.25.1
httpcore==0.18.0
flask_session
Flask-Session>=0.4.0
Flask
requests
flask_cors
docx
python-docx
pyyaml
fastapi
uvicorn
python-dotenv
redis
uvicorn[standard]
transformers
google-cloud-storage
